update-v2.2

+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/MPADao.java
@@ -3,11 +3,10 @@ package ru.yandex.practicum.filmorate.storage.dao;
 import ru.yandex.practicum.filmorate.models.MPARating;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface MPADao {
 
-    MPARating getById(long id);
+    MPARating getById(int id);
 
     List<MPARating> getAll();
